+++ packages/sui-react-initial-props/src/index.js
@@ -1,4 +1,10 @@
-export createClientContextFactoryParams from './createClientContextFactoryParams'
-export createServerContextFactoryParams from './createServerContextFactoryParams'
-export loadPage from './loadPage'
-export ssrComponentWithInitialProps from './ssrComponentWithInitialProps'
+export {
+  default as createClientContextFactoryParams
+} from './createClientContextFactoryParams'
+export {
+  default as createServerContextFactoryParams
+} from './createServerContextFactoryParams'
+export {default as loadPage} from './loadPage'
+export {
+  default as ssrComponentWithInitialProps
+} from './ssrComponentWithInitialProps'