+++ packages/sui-studio/workbench/src/components/Root/index.js
@@ -1,7 +1,6 @@
 /* eslint import/no-webpack-loader-syntax:0 */
 import React from 'react'
 import PropTypes from 'prop-types'
-import {hot} from 'react-hot-loader'
 
 import Header from '../Header'
 import Select from '../Select'
@@ -94,4 +93,4 @@ class Root extends React.PureComponent {
   }
 }
 
-export default hot(module)(Root)
+export default Root