+++ packages/sui-domain/CHANGELOG.md
@@ -2,6 +2,36 @@
 
 All notable changes to this project will be documented in this file.
 
+<a name="1.16.0"></a>
+# 1.16.0 (2019-03-18)
+
+
+### Bug Fixes
+
+* Release library with correct babel-preset-sui installed ([be9486f](https://github.com/SUI-Components/sui/commit/be9486f))
+
+
+
+<a name="1.15.0"></a>
+# 1.15.0 (2019-03-18)
+
+
+### Features
+
+* Move to babel-preset-sui@3 ([84a3a57](https://github.com/SUI-Components/sui/commit/84a3a57))
+
+
+
+<a name="1.14.0"></a>
+# 1.14.0 (2019-03-13)
+
+
+### Bug Fixes
+
+* avoid using not supported export from webpack ([566abb5](https://github.com/SUI-Components/sui/commit/566abb5))
+
+
+
 <a name="1.13.0"></a>
 # 1.13.0 (2019-02-19)