+++ packages/sui-react-context/CHANGELOG.md
@@ -2,3 +2,23 @@
 
 All notable changes to this project will be documented in this file.
 
+<a name="1.4.0"></a>
+# 1.4.0 (2019-03-18)
+
+
+### Bug Fixes
+
+* Release library with correct babel-preset-sui installed ([807e79c](https://github.com/SUI-Components/sui/commit/807e79c))
+
+
+
+<a name="1.3.0"></a>
+# 1.3.0 (2019-03-18)
+
+
+### Features
+
+* Move to babel-preset-sui@3 ([b43baca](https://github.com/SUI-Components/sui/commit/b43baca))
+
+
+