+++ packages/sui-studio-utils/CHANGELOG.md
@@ -2,6 +2,36 @@
 
 All notable changes to this project will be documented in this file.
 
+<a name="1.6.0"></a>
+# 1.6.0 (2019-03-19)
+
+
+### Bug Fixes
+
+* update export syntax ([a510111](https://github.com/SUI-Components/sui/commit/a510111))
+
+
+
+<a name="1.5.0"></a>
+# 1.5.0 (2019-03-18)
+
+
+### Bug Fixes
+
+* Release library with correct babel-preset-sui installed ([7b16e5a](https://github.com/SUI-Components/sui/commit/7b16e5a))
+
+
+
+<a name="1.4.0"></a>
+# 1.4.0 (2019-03-18)
+
+
+### Features
+
+* Move to babel-preset-sui@3 ([90761a0](https://github.com/SUI-Components/sui/commit/90761a0))
+
+
+
 <a name="1.3.0"></a>
 # 1.3.0 (2019-02-19)